% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{handle_ipynb}
\alias{handle_ipynb}
\title{Handle Jupyter notebooks (.ipynb) by extracting code + markdown cell sources}
\usage{
handle_ipynb(path, max_chars = 16000)
}
\description{
Handle Jupyter notebooks (.ipynb) by extracting code + markdown cell sources
}
