% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{handle_rdata}
\alias{handle_rdata}
\title{Inspect .RData/.rda files and return object summary (do not attach to global env)}
\usage{
handle_rdata(path, max_chars = 2000)
}
\description{
Inspect .RData/.rda files and return object summary (do not attach to global env)
}
