% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{handle_rds}
\alias{handle_rds}
\title{Inspect .rds by reading object metadata (class, length) but avoid printing full object}
\usage{
handle_rds(path, max_chars = 2000)
}
\description{
Inspect .rds by reading object metadata (class, length) but avoid printing full object
}
