% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{make_ellmer_chat_fn}
\alias{make_ellmer_chat_fn}
\title{Create chat wrapper for an ellmer chat object or similar}
\usage{
make_ellmer_chat_fn(chat_obj, response_field = NULL)
}
\description{
Create chat wrapper for an ellmer chat object or similar
}
