% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{scan_project_files}
\alias{scan_project_files}
\title{Scan project for files to document (honoring excludes)}
\usage{
scan_project_files(
  path = ".",
  include = c("R", "r", "Rmd", "rmd", "qmd", "qmarkdown", "md", "markdown", "html",
    "htm", "ipynb", "pdf", "docx", "doc", "pptx", "rds", "rda", "rdata", "csv", "sql",
    "json", "yml", "yaml"),
  exclude_dirs = c("packrat", "renv", "node_modules", ".git", "demo"),
  exclude_patterns = c("\\\\.Rproj$", "^README\\\\.md$")
)
}
\description{
Scan project for files to document (honoring excludes)
}
